refactored code and added outro timing in episode source
